Add POST and DELETE functionality
Included an endpoint for handling POST and DELETE request.
Modified frontEnd "add" and "delete" handler to handle them.
This enable new todo's to be added and saved on the backend and includes
on the response.

NOTE: Data is stored in memory on backend server for now. Thus, its not
persistent (lost when server restarts)