[i] Add middleware for our JWT authentication
This commit introduces a new middleware function 'requireAuth',
which is responsible for verifying user authentication using JWT.

The middleware checks for Authentication header and verifies
the JWT token. We then extract the '_id' from it.

Then we attached the user's _id to the request object (user) that further
carries on to other API request.

If fail, we do not proceed further as we throw error, avoiding 'next()'
to be invoked. So any GET, POST, etc., will not run after it.

Overall this feature enhances security and access control for authenticated
routes.