! - Implementing Authorized requests
Goal of the commit is to have any API requests to be authorized before
performing any CRUD operation. Only once a user logged in, they get auth token
that enables them to proceed with the operations.

Backend:
Some changes needed on our server to correctly handles API request and
CORS for different endpoints.
OPTIONS route handler added for this to handle preflight request.

TODO: use some 'cors' middleware to better sort out CORS issue.

FrontEnd:
react 'context & hooks' for auth is now used to get logged in user's
token. This token is now use as "Authentication" header for any API CRUD request.

With this, `GetServerSideProps` is no longer used to fetch initial data
(list of todos) as it cant accept react hooks and context due to it runs before the
component even mounts

TODO: Alert box is used for now to warn if no 'user'token/session when
performing request or actions. We can improve with proper error/warning message on
the GUI.